CoinDetailsScreen implemented. Code refactored.

+++ app/src/main/java/com/spinner/cryptocurrencies/data/remote/dto/ConvertedLast.kt
@@ -1,10 +1,8 @@
 package com.spinner.cryptocurrencies.data.remote.dto
 
 
-import com.google.gson.annotations.SerializedName
-
 data class ConvertedLast(
     val btc: Double,
     val eth: Double,
-    val usd: Int
+    val usd: Double,
 )
