+++ app/src/main/java/com/spinner/cryptocurrencies/data/remote/dto/ConvertedVolume.kt
@@ -1,10 +1,8 @@
 package com.spinner.cryptocurrencies.data.remote.dto
 
 
-import com.google.gson.annotations.SerializedName
-
 data class ConvertedVolume(
-    val btc: Int,
-    val eth: Int,
-    val usd: Int
+    val btc: Double,
+    val eth: Double,
+    val usd: Double,
 )
